move step1에 썼던 클래스들 패키지 변경
