+++ src/main/java/step1/Calculator.java
@@ -1,6 +1,4 @@
-package utils;
-
-import domain.Expression;
+package step1;
 
 public class Calculator {
     public static int calculate(String input) {
