+++ src/main/java/step1/Operator.java
@@ -1,4 +1,4 @@
-package domain;
+package step1;
 
 import java.util.Arrays;
