+++ src/test/java/step1/CalculatorTest.java
@@ -1,4 +1,4 @@
-package utils;
+package step1;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
