+++ src/test/java/step1/ExpressionTest.java
@@ -1,8 +1,7 @@
-package domain;
+package step1;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import utils.Calculator;
 
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
