+++ src/test/java/step1/OperatorTest.java
@@ -1,4 +1,4 @@
-package domain;
+package step1;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
